+++ src/EventStore.ClientAPI/Internal/EventStoreNodeConnection.cs
@@ -86,12 +86,12 @@ namespace EventStore.ClientAPI.Internal
             _handler.EnqueueMessage(new CloseConnectionMessage("Connection close requested by client.", null));
         }
 
-        public Task<DeleteResult> DeleteStreamAsync(string stream, int expectedVersion, UserCredentials userCredentials = null)
+        public Task<DeleteResult> DeleteStreamAsync(string stream, long expectedVersion, UserCredentials userCredentials = null)
         {
             return DeleteStreamAsync(stream, expectedVersion, false, userCredentials);
         }
 
-        public Task<DeleteResult> DeleteStreamAsync(string stream, int expectedVersion, bool hardDelete, UserCredentials userCredentials = null)
+        public Task<DeleteResult> DeleteStreamAsync(string stream, long expectedVersion, bool hardDelete, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
 
@@ -101,7 +101,7 @@ namespace EventStore.ClientAPI.Internal
             return source.Task;
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, params EventData[] events)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, params EventData[] events)
         {
 // ReSharper disable RedundantArgumentDefaultValue
 // ReSharper disable RedundantCast
@@ -110,14 +110,14 @@ namespace EventStore.ClientAPI.Internal
 // ReSharper restore RedundantArgumentDefaultValue
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, UserCredentials userCredentials, params EventData[] events)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, UserCredentials userCredentials, params EventData[] events)
         {
 // ReSharper disable RedundantCast
             return AppendToStreamAsync(stream, expectedVersion, (IEnumerable<EventData>)events, userCredentials);
 // ReSharper restore RedundantCast
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, IEnumerable<EventData> events, UserCredentials userCredentials = null)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, IEnumerable<EventData> events, UserCredentials userCredentials = null)
         {
 // ReSharper disable PossibleMultipleEnumeration
             Ensure.NotNullOrEmpty(stream, "stream");
@@ -130,7 +130,7 @@ namespace EventStore.ClientAPI.Internal
 // ReSharper restore PossibleMultipleEnumeration
         }
 
-        public Task<ConditionalWriteResult> ConditionalAppendToStreamAsync(string stream, int expectedVersion, IEnumerable<EventData> events,
+        public Task<ConditionalWriteResult> ConditionalAppendToStreamAsync(string stream, long expectedVersion, IEnumerable<EventData> events,
             UserCredentials userCredentials = null)
         {
             // ReSharper disable PossibleMultipleEnumeration
@@ -144,7 +144,7 @@ namespace EventStore.ClientAPI.Internal
             // ReSharper restore PossibleMultipleEnumeration
         }
 
-        public Task<EventStoreTransaction> StartTransactionAsync(string stream, int expectedVersion, UserCredentials userCredentials = null)
+        public Task<EventStoreTransaction> StartTransactionAsync(string stream, long expectedVersion, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
 
@@ -184,7 +184,7 @@ namespace EventStore.ClientAPI.Internal
         }
 
 
-        public Task<EventReadResult> ReadEventAsync(string stream, int eventNumber, bool resolveLinkTos, UserCredentials userCredentials = null)
+        public Task<EventReadResult> ReadEventAsync(string stream, long eventNumber, bool resolveLinkTos, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
             if (eventNumber < -1) throw new ArgumentOutOfRangeException("eventNumber");
@@ -195,7 +195,7 @@ namespace EventStore.ClientAPI.Internal
             return source.Task;
         }
 
-        public Task<StreamEventsSlice> ReadStreamEventsForwardAsync(string stream, int start, int count, bool resolveLinkTos, UserCredentials userCredentials = null)
+        public Task<StreamEventsSlice> ReadStreamEventsForwardAsync(string stream, long start, int count, bool resolveLinkTos, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
             Ensure.Nonnegative(start, "start");
@@ -208,7 +208,7 @@ namespace EventStore.ClientAPI.Internal
             return source.Task;
         }
 
-        public Task<StreamEventsSlice> ReadStreamEventsBackwardAsync(string stream, int start, int count, bool resolveLinkTos, UserCredentials userCredentials = null)
+        public Task<StreamEventsSlice> ReadStreamEventsBackwardAsync(string stream, long start, int count, bool resolveLinkTos, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
             Ensure.Positive(count, "count");
@@ -269,7 +269,7 @@ namespace EventStore.ClientAPI.Internal
         }
 
         public EventStoreStreamCatchUpSubscription SubscribeToStreamFrom(string stream,
-                                                                         int? lastCheckpoint,
+                                                                         long? lastCheckpoint,
                                                                          bool resolveLinkTos,
                                                                          Action<EventStoreCatchUpSubscription, ResolvedEvent> eventAppeared,
                                                                          Action<EventStoreCatchUpSubscription> liveProcessingStarted = null,
@@ -284,7 +284,7 @@ namespace EventStore.ClientAPI.Internal
 
         public EventStoreStreamCatchUpSubscription SubscribeToStreamFrom(
                 string stream,
-                int? lastCheckpoint,
+                long? lastCheckpoint,
                 CatchUpSubscriptionSettings settings,
                 Action<EventStoreCatchUpSubscription, ResolvedEvent> eventAppeared,
                 Action<EventStoreCatchUpSubscription> liveProcessingStarted = null,
@@ -459,12 +459,12 @@ namespace EventStore.ClientAPI.Internal
 
 */
 
-        public Task<WriteResult> SetStreamMetadataAsync(string stream, int expectedMetastreamVersion, StreamMetadata metadata, UserCredentials userCredentials = null)
+        public Task<WriteResult> SetStreamMetadataAsync(string stream, long expectedMetastreamVersion, StreamMetadata metadata, UserCredentials userCredentials = null)
         {
             return SetStreamMetadataAsync(stream, expectedMetastreamVersion, metadata.AsJsonBytes(), userCredentials);
         }
 
-        public Task<WriteResult> SetStreamMetadataAsync(string stream, int expectedMetastreamVersion, byte[] metadata, UserCredentials userCredentials = null)
+        public Task<WriteResult> SetStreamMetadataAsync(string stream, long expectedMetastreamVersion, byte[] metadata, UserCredentials userCredentials = null)
         {
             Ensure.NotNullOrEmpty(stream, "stream");
             if (SystemStreams.IsMetastream(stream))
@@ -516,7 +516,7 @@ namespace EventStore.ClientAPI.Internal
                     case EventReadStatus.NoStream:
                         return new RawStreamMetadataResult(stream, false, -1, Empty.ByteArray);
                     case EventReadStatus.StreamDeleted:
-                        return new RawStreamMetadataResult(stream, true, int.MaxValue, Empty.ByteArray);
+                        return new RawStreamMetadataResult(stream, true, long.MaxValue, Empty.ByteArray);
                     default:
                         throw new ArgumentOutOfRangeException(string.Format("Unexpected ReadEventResult: {0}.", res.Status));
                 }
