+++ src/EventStore.Core.Tests/ClientAPI/catch_up_subscription_handles_errors.cs
@@ -439,7 +439,7 @@ namespace EventStore.Core.Tests.ClientAPI
     internal class FakeEventStoreConnection : IEventStoreConnection
     {
         private Func<Position, int, bool, UserCredentials, Task<AllEventsSlice>> _readAllEventsForwardAsync;
-        private Func<string, int, int, Task<StreamEventsSlice>> _readStreamEventsForwardAsync;
+        private Func<string, long, int, Task<StreamEventsSlice>> _readStreamEventsForwardAsync;
         private Func<string, Action<EventStoreSubscription, ResolvedEvent>, Action<EventStoreSubscription, SubscriptionDropReason, Exception>, Task<EventStoreSubscription>> _subscribeToStreamAsync;
 
         public void Dispose()
@@ -460,38 +460,38 @@ namespace EventStore.Core.Tests.ClientAPI
             throw new NotImplementedException();
         }
 
-        public Task<DeleteResult> DeleteStreamAsync(string stream, int expectedVersion, UserCredentials userCredentials = null)
+        public Task<DeleteResult> DeleteStreamAsync(string stream, long expectedVersion, UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public Task<DeleteResult> DeleteStreamAsync(string stream, int expectedVersion, bool hardDelete, UserCredentials userCredentials = null)
+        public Task<DeleteResult> DeleteStreamAsync(string stream, long expectedVersion, bool hardDelete, UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, params EventData[] events)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, params EventData[] events)
         {
             throw new NotImplementedException();
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, UserCredentials userCredentials, params EventData[] events)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, UserCredentials userCredentials, params EventData[] events)
         {
             throw new NotImplementedException();
         }
 
-        public Task<WriteResult> AppendToStreamAsync(string stream, int expectedVersion, IEnumerable<EventData> events, UserCredentials userCredentials = null)
+        public Task<WriteResult> AppendToStreamAsync(string stream, long expectedVersion, IEnumerable<EventData> events, UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public Task<ConditionalWriteResult> ConditionalAppendToStreamAsync(string stream, int expectedVersion, IEnumerable<EventData> events,
+        public Task<ConditionalWriteResult> ConditionalAppendToStreamAsync(string stream, long expectedVersion, IEnumerable<EventData> events,
             UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public Task<EventStoreTransaction> StartTransactionAsync(string stream, int expectedVersion, UserCredentials userCredentials = null)
+        public Task<EventStoreTransaction> StartTransactionAsync(string stream, long expectedVersion, UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
@@ -501,23 +501,23 @@ namespace EventStore.Core.Tests.ClientAPI
             throw new NotImplementedException();
         }
 
-        public Task<EventReadResult> ReadEventAsync(string stream, int eventNumber, bool resolveLinkTos, UserCredentials userCredentials = null)
+        public Task<EventReadResult> ReadEventAsync(string stream, long eventNumber, bool resolveLinkTos, UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public void HandleReadStreamEventsForwardAsync(Func<string, int, int, Task<StreamEventsSlice>> callback)
+        public void HandleReadStreamEventsForwardAsync(Func<string, long, int, Task<StreamEventsSlice>> callback)
         {
             _readStreamEventsForwardAsync = callback;
         }
 
-        public Task<StreamEventsSlice> ReadStreamEventsForwardAsync(string stream, int start, int count, bool resolveLinkTos,
+        public Task<StreamEventsSlice> ReadStreamEventsForwardAsync(string stream, long start, int count, bool resolveLinkTos,
             UserCredentials userCredentials = null)
         {
             return _readStreamEventsForwardAsync(stream, start, count);
         }
 
-        public Task<StreamEventsSlice> ReadStreamEventsBackwardAsync(string stream, int start, int count, bool resolveLinkTos,
+        public Task<StreamEventsSlice> ReadStreamEventsBackwardAsync(string stream, long start, int count, bool resolveLinkTos,
             UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
@@ -551,7 +551,7 @@ namespace EventStore.Core.Tests.ClientAPI
             return _subscribeToStreamAsync(stream, eventAppeared, subscriptionDropped);
         }
 
-        public EventStoreStreamCatchUpSubscription SubscribeToStreamFrom(string stream, int? lastCheckpoint, bool resolveLinkTos,
+        public EventStoreStreamCatchUpSubscription SubscribeToStreamFrom(string stream, long? lastCheckpoint, bool resolveLinkTos,
             Action<EventStoreCatchUpSubscription, ResolvedEvent> eventAppeared, Action<EventStoreCatchUpSubscription> liveProcessingStarted = null, Action<EventStoreCatchUpSubscription, SubscriptionDropReason, Exception> subscriptionDropped = null,
             UserCredentials userCredentials = null, int readBatchSize = 500)
         {
@@ -560,7 +560,7 @@ namespace EventStore.Core.Tests.ClientAPI
 
         public EventStoreStreamCatchUpSubscription SubscribeToStreamFrom(
             string stream,
-            int? lastCheckpoint,
+            long? lastCheckpoint,
             CatchUpSubscriptionSettings settings,
             Action<EventStoreCatchUpSubscription, ResolvedEvent> eventAppeared,
             Action<EventStoreCatchUpSubscription> liveProcessingStarted = null,
@@ -618,13 +618,13 @@ namespace EventStore.Core.Tests.ClientAPI
             throw new NotImplementedException();
         }
 
-        public Task<WriteResult> SetStreamMetadataAsync(string stream, int expectedMetastreamVersion, StreamMetadata metadata,
+        public Task<WriteResult> SetStreamMetadataAsync(string stream, long expectedMetastreamVersion, StreamMetadata metadata,
             UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
         }
 
-        public Task<WriteResult> SetStreamMetadataAsync(string stream, int expectedMetastreamVersion, byte[] metadata,
+        public Task<WriteResult> SetStreamMetadataAsync(string stream, long expectedMetastreamVersion, byte[] metadata,
             UserCredentials userCredentials = null)
         {
             throw new NotImplementedException();
