+++ src/EventStore.Core/Services/PersistentSubscription/IPersistentSubscriptionCheckpointWriter.cs
@@ -4,7 +4,7 @@ namespace EventStore.Core.Services.PersistentSubscription
 {
     public interface IPersistentSubscriptionCheckpointWriter
     {
-        void BeginWriteState(int state);
+        void BeginWriteState(long state);
         void BeginDelete(Action<IPersistentSubscriptionCheckpointWriter> completed);
     }
 }
